Craft demo

+++ src/main/java/com/thetechmaddy/ecommerce/controllers/OrdersController.java
@@ -1,11 +1,11 @@
 package com.thetechmaddy.ecommerce.controllers;
 
 import com.fasterxml.jackson.annotation.JsonView;
+import com.thetechmaddy.ecommerce.domains.carts.Cart;
+import com.thetechmaddy.ecommerce.domains.carts.CartItem;
 import com.thetechmaddy.ecommerce.domains.orders.Order;
-import com.thetechmaddy.ecommerce.models.AppConstants;
-import com.thetechmaddy.ecommerce.models.JsonViews;
-import com.thetechmaddy.ecommerce.models.OrderStatus;
-import com.thetechmaddy.ecommerce.models.OrderSummary;
+import com.thetechmaddy.ecommerce.exceptions.UnsupportedBusinessOperationException;
+import com.thetechmaddy.ecommerce.models.*;
 import com.thetechmaddy.ecommerce.models.delivery.DeliveryInfo;
 import com.thetechmaddy.ecommerce.models.filters.OrderFilters;
 import com.thetechmaddy.ecommerce.models.payments.PaymentInfo;
@@ -13,6 +13,7 @@ import com.thetechmaddy.ecommerce.models.requests.CognitoUser;
 import com.thetechmaddy.ecommerce.models.requests.OrderRequest;
 import com.thetechmaddy.ecommerce.models.responses.ApiResponse;
 import com.thetechmaddy.ecommerce.models.responses.Paged;
+import com.thetechmaddy.ecommerce.services.CartsService;
 import com.thetechmaddy.ecommerce.services.OrdersService;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import jakarta.validation.Valid;
@@ -20,7 +21,11 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
+
 import static com.thetechmaddy.ecommerce.models.AppConstants.CURRENT_USER_REQUEST_ATTRIBUTE;
+import static com.thetechmaddy.ecommerce.models.AppConstants.DEFAULT_DELETE_ORDER_STATUS;
+import static com.thetechmaddy.ecommerce.models.OrderStatus.PENDING;
 
 @RestController
 @RequestMapping("/api/orders")
@@ -29,19 +34,31 @@ import static com.thetechmaddy.ecommerce.models.AppConstants.CURRENT_USER_REQUES
 public class OrdersController extends BaseController {
 
     private final OrdersService ordersService;
+    private final CartsService cartsService;
 
     @PostMapping("/initiate")
     @JsonView({JsonViews.OrderInitiateResponse.class})
     public ApiResponse<Order> initiateOrder(@RequestAttribute(name = CURRENT_USER_REQUEST_ATTRIBUTE) CognitoUser cognitoUser,
                                             @RequestBody @Valid OrderRequest orderRequest) {
+        String userId = cognitoUser.getCognitoSub();
 
-        Order pendingOrder = ordersService.getUserOrderInPendingStatus(cognitoUser.getCognitoSub());
+        Cart cart = cartsService.getCart(orderRequest.getCartId(), userId);
+        List<CartItem> cartItems = cart.getCartItems();
 
-        if (pendingOrder == null) {
-            pendingOrder = ordersService.createNewOrder(cognitoUser.getCognitoSub(), orderRequest);
+        Order order = ordersService.getUserOrderInPendingStatus(userId);
+        if (order == null) {
+            return ApiResponse.success(ordersService.createNewOrder(userId, orderRequest));
         }
 
-        return ApiResponse.success(pendingOrder);
+        CartItemsOrderItemsMatcher itemsMatcher = new CartItemsOrderItemsMatcher(cartItems, order.getOrderItems());
+        if (itemsMatcher.matches()) {
+            return ApiResponse.success(order);
+        }
+
+        ordersService.deletePendingOrder(order.getId(), userId);
+        cartsService.lockCart(cart.getId(), userId);
+        return ApiResponse.success(ordersService.createNewOrder(userId, orderRequest));
+
     }
 
     @PutMapping("/{orderId}/payment")
@@ -91,8 +108,14 @@ public class OrdersController extends BaseController {
     }
 
     @DeleteMapping("/{orderId}")
-    public ApiResponse<?> deletePendingOrder(@RequestAttribute(name = CURRENT_USER_REQUEST_ATTRIBUTE) CognitoUser cognitoUser,
-                                             @PathVariable("orderId") long orderId) {
+    public ApiResponse<?> deleteOrder(@RequestAttribute(name = CURRENT_USER_REQUEST_ATTRIBUTE) CognitoUser cognitoUser,
+                                      @PathVariable("orderId") long orderId,
+                                      @RequestParam(name = "status", required = false,
+                                              defaultValue = DEFAULT_DELETE_ORDER_STATUS) String deleteOrderStatus) {
+        if (OrderStatus.parse(deleteOrderStatus) != PENDING) {
+            throw new UnsupportedBusinessOperationException("Delete order other than PENDING status not supported");
+        }
+
         ordersService.deletePendingOrder(orderId, cognitoUser.getCognitoSub());
         return ApiResponse.success();
     }
